+++ packages/create-keystone-app/example-projects/blank/index.js
@@ -1,14 +1,14 @@
-const { Keystone } = require('@keystone-alpha/keystone');
-const { GraphQLApp } = require('@keystone-alpha/app-graphql');
-const { AdminUIApp } = require('@keystone-alpha/app-admin-ui');
+const { Keystone } = require('@keystonejs/keystone');
+const { GraphQLApp } = require('@keystonejs/app-graphql');
+const { AdminUIApp } = require('@keystonejs/app-admin-ui');
 /* keystone-cli: generated-code */
-const { MongooseAdapter: Adapter } = require('@keystone-alpha/adapter-mongoose');
+const { MongooseAdapter: Adapter } = require('@keystonejs/adapter-mongoose');
 const PROJECT_NAME = 'My KeystoneJS Project';
 /* /keystone-cli: generated-code */
 
 /**
  * You've got a new KeystoneJS Project! Things you might want to do next:
- * - Add adapter config options (See: https://keystonejs.com/keystone-alpha/adapter-mongoose/)
+ * - Add adapter config options (See: https://keystonejs.com/keystonejs/adapter-mongoose/)
  * - Select configure access control and authentication (See: https://keystonejs.com/api/access-control)
  */